requirements.txt
flask==3.0.0
gitpython==3.1.31
sqlalchemy==1.4.41
aiofiles==23.1.0
